testHelper auf mehrere Dateien  aufgesplittet
