+++ base/api/testHelper.page.ts
@@ -1,0 +1,7 @@
+export interface UserInfo {
+    username: string,
+    password: string,
+    rolleId: string,
+    organisationId: string,
+    personId: string;
+}
